src/main/java/ro/dix/algorithms/string/StringReverse.java
package ro.dix.algorithms.string;

/**
 *
 * @author dan
 */
public class StringReverse {

    public String reverse(String source) {
        int i, len = source.length();
        StringBuilder dest = new StringBuilder(len);
        for (i = (len - 1); i >= 0; i--) {
            dest.append(source.charAt(i));
        }
        return dest.toString();
    }

    public String reverseInline(String s) {
        char[] chars = s.toCharArray();
        for (int i = 0, j = chars.length - 1; i < j; ++i, --j) {
            char tmp = chars[i];
            chars[i] = chars[j];
            chars[j] = tmp;
        }
        return new String(chars);
    }
}
